LayoutTests/imported/w3c/web-platform-tests/xhr/send-no-response-event-order-expected.txt


FAIL XMLHttpRequest: The send() method: event order when there is no response entity body assert_equals: expected "progress(0,0,false)" but got "readystatechange(4)"
